+++ tp-orgDeDatos/Pruebas/main.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-//const long megaByte = 1024*1024;
+//const int megaByte = 1024*1024;
 const unsigned char orden = 3;
 
 int main(int argc,char *argv[]){
@@ -19,7 +19,7 @@ int main(int argc,char *argv[]){
 	if (*operacionARealizar == 'c'){
 		FileManager<char> fileManager;
 		char* buffer = fileManager.processFile(nombreDelArchivo);
-		unsigned long size = fileManager.getSize(nombreDelArchivo);
+		unsigned int size = fileManager.getSize(nombreDelArchivo);
 
 		cout << "El archivo ha sido abierto y colocado en memoria correctamente" << endl;
 
@@ -29,12 +29,12 @@ int main(int argc,char *argv[]){
 	else if(*operacionARealizar=='d'){
 //		FileManager<unsigned short> fileManager;
 //		unsigned short* buffer = fileManager.processFile(nombreDelArchivo);
-//		unsigned long size = fileManager.getSize(nombreDelArchivo);
+//		unsigned int size = fileManager.getSize(nombreDelArchivo);
 
 		cout << "El archivo ha sido abierto y colocado en memoria correctamente" << endl;
 
 		unsigned short buffer[24] = {97,98,99,100,101,0,0,4,0,0,4,0,112,1,0,2,0,0,0,1,0,0,0,0};
-		unsigned long size = 24;
+		unsigned int size = 24;
 		compresor.descomprimir(buffer,size);
 //		free(buffer);
 	}
